Add Features
This is a big progress because of the following:
1. Create metadata object for UI to retrieve all the info of datafiles
from logic
2. Create API for UI to delete datafiles/indexing Entity
3. Create Auto-Update Indexing System Mechanism such that it would
update the indexing system Every Certain Amount of Interval or Outsider
trigger this mechanism
4. Create Auto Deletion mechanism for Expired Files, datafiles and
entries which is older than certain amount of time will be deleted
automatically. And this checking is done in every fixed amount of
Interval (To test, I set it to every minute)
5. Refactor and Code Cleanning for JavaCC parser
What to be done next?
Create Unit Testing for all these mechanism and Implement the Searching
feature.