Connect Everything Together
Connect Every Component of Back End Logic together so that it supports
the following by running at main:

1. UI component is now able to pass the userName, userEmail, the number
of display entries, the user sync directory and the expired time in
hours to logic

2. System is able to generate the unique local key for non-login default
user in the current machine

3. System is able to generate the unique key for a particular user

4. System is able to handle different configuration data generation
between new user in the current machine and the old user in this machine

5. System is able to encrypt the data file with the user unique key
instead of fixed key

6. System is able to decrypt the data file with correct user Key.

7. System is able to know whether a user is new to the System by
checking whether his/hers key exists in the key set.

Next thing to do:
Cleanning + Testing. I am almost finished!