Lucene Testing
Bug Fix at previous Security Test
Bug found and Fix and document deletion.
Previously file name of data files will be stored as a StringField in
lucene document to preserve the integrity of the filename. As
StringField in lucene is not tokenized, it will be case sensitive. In
fact, all previous deletion query will be parsed in such a way that all
the terms within the query will be tokenized and converted to small
case. As such, deletion will fail in case the data file name consists of
any capital characters. Fix the bug by adding a SearchFilename field in
the original document design which will be the tokenized format of the
filename.

Implement testing for lucene indexing system to test the insertion,
deletion and search functionality

Create Continuous Integrations that binds to this project

Robustness  Programming
Improved on performance by changing the mechanism of transmission of
data, which is the bottleneck for the software according to profiling
result.